add LOCALLOG_AUTO_DELETE env

+++ hack/test-logcollect.sh
@@ -118,8 +118,10 @@ fi
 "${GRS_ROOT}/hack/test-loganalysis.sh"
 echo "Copying logs to destination instance."
 copy-logs "${DES_LOG_INSTANCE}" "${DES_LOG_INSTANCE_ZONE}" "${DESTINATION}" "${DES_LOG_DIR}"
-echo "Removing local copy from ${DESTINATION}"
-sudo rm -r "${DESTINATION}"
+if [ "${LOCALLOG_AUTO_DELETE}" == "true" ]; then
+        echo "Removing local copy from ${DESTINATION}"
+        sudo rm -r "${DESTINATION}"
+fi
